Get right recent events showing up (fix order of enhancers in createStore)
