+++ functions/src/actionRunner/index.js
@@ -1,31 +1,12 @@
-import { invoke } from 'lodash'
-import { updateResponseOnRTDB, updateRequestAsStarted } from './utils'
+import * as functions from 'firebase-functions'
+import {
+  updateResponseOnRTDB,
+  updateRequestAsStarted,
+  writeProjectEvent
+} from './utils'
 import { ACTION_RUNNER_REQUESTS_PATH } from './constants'
-import { runStepsFromEvent } from './steps'
+import { runStepsFromEvent } from './stepsRunner'
 import { to } from '../utils/async'
-import * as admin from 'firebase-admin'
-const functions = require('firebase-functions')
-
-async function writeProjectEvent(projectId, extraEventAttributes) {
-  const eventObject = {
-    createdByType: 'system',
-    createdAt: admin.firestore.FieldValue.serverTimestamp(),
-    ...extraEventAttributes
-  }
-  const eventsRef = admin
-    .firestore()
-    .collection('projects')
-    .doc(projectId)
-    .collection('events')
-  const [addErr, addRes] = await to(eventsRef.add(eventObject))
-  if (addErr) {
-    const errMsg = `Error adding event data to Project events for project: ${projectId}`
-    console.error(errMsg)
-    console.error(addErr.message || addErr)
-    throw new Error(errMsg)
-  }
-  return addRes
-}
 
 /**
  * @name actionRunner
@@ -38,18 +19,23 @@ export default functions.database
   .onCreate(runAction)
 
 async function runAction(event) {
-  const eventData = invoke(event.data, 'val') || {}
+  const eventData = event.data && event.data.val ? event.data.val() : {}
   console.log('Action run request recieved', eventData)
+
+  // Running an action not supported without projectId
   if (!eventData.projectId) {
     throw new Error('projectId parameter is required')
   }
+
   const startEvent = {
     eventType: 'startActionRun',
     eventData
   }
-
   await updateRequestAsStarted(event)
   await writeProjectEvent(eventData.projectId, startEvent)
+  // console.log('Start event sent successfully. Checking for backups...')
+  // const [err, result] = await to(runStepsFromEvent(event))
+
   console.log('Start event sent successfully. Starting run of steps...')
   const [err, result] = await to(runStepsFromEvent(event))
   if (err) {
@@ -64,10 +50,13 @@ async function runAction(event) {
     await writeProjectEvent(eventData.projectId, errorEvent)
     throw err
   }
+
   console.log('Action completed writing event to project...')
+
   // Write event to project indicating action run is complete
   const finishedEvent = { eventType: 'finishActionRun', eventData }
   await writeProjectEvent(eventData.projectId, finishedEvent)
+
   console.log('Action completed successfully!')
   return result
 }
